requirements.txt
nltk
afinn
pandas
python-dotenv
requests
fastapi==0.109.0
uvicorn==0.25.0
pydantic==1.10.7
sqlparse==0.4.4
pytest==7.4.0
textstat==0.7.2
textblob==0.17.1
nltk==3.8.1
readability-lxml==0.8.1
streamlit
beautifulsoup4
googlesearch-python
transformers
torch
numpy
chromadb
fpdf2